+++ mm2src/mm2_main/src/lp_stats.rs
@@ -7,7 +7,7 @@ use futures::lock::Mutex as AsyncMutex;
 use http::StatusCode;
 use mm2_core::mm_ctx::{from_ctx, MmArc};
 use mm2_err_handle::prelude::*;
-use mm2_libp2p::application::request_response::network_info::PeerInfoRequest;
+use mm2_libp2p::application::request_response::peer_info::PeerInfoRequest;
 use mm2_libp2p::{encode_message, NetworkInfo, PeerId, RelayAddress, RelayAddressError};
 use serde_json::{self as json, Value as Json};
 use std::collections::{HashMap, HashSet};
