+++ mm2src/mm2_p2p/src/application/request_response/peer_info.rs
